+++ container/raw/handler.go
@@ -26,7 +26,7 @@ import (
 	"github.com/google/cadvisor/container/libcontainer"
 	"github.com/google/cadvisor/fs"
 	info "github.com/google/cadvisor/info/v1"
-	"github.com/google/cadvisor/utils"
+	"github.com/google/cadvisor/machine"
 
 	"github.com/golang/glog"
 	"github.com/opencontainers/runc/libcontainer/cgroups"
@@ -63,30 +63,6 @@ type rawContainerHandler struct {
 	ignoreMetrics container.MetricSet
 
 	pid int
-}
-
-func (self *rawContainerHandler) GetCgroupPaths() map[string]string {
-	return self.cgroupPaths
-}
-
-func (self *rawContainerHandler) GetMachineInfoFactory() info.MachineInfoFactory {
-	return self.machineInfoFactory
-}
-
-func (self *rawContainerHandler) GetName() string {
-	return self.name
-}
-
-func (self *rawContainerHandler) GetExternalMounts() []common.Mount {
-	return self.externalMounts
-}
-
-func (self *rawContainerHandler) HasNetwork() bool {
-	return false
-}
-
-func (self *rawContainerHandler) HasFilesystem() bool {
-	return false
 }
 
 func isRootCgroup(name string) bool {
@@ -164,7 +140,41 @@ func (self *rawContainerHandler) Start() {}
 func (self *rawContainerHandler) Cleanup() {}
 
 func (self *rawContainerHandler) GetSpec() (info.ContainerSpec, error) {
-	return common.GetSpec(self)
+	const hasNetwork = false
+	hasFilesystem := isRootCgroup(self.name) || len(self.externalMounts) > 0
+	spec, err := common.GetSpec(self.cgroupPaths, self.machineInfoFactory, hasNetwork, hasFilesystem)
+	if err != nil {
+		return spec, err
+	}
+
+	if isRootCgroup(self.name) {
+		// Check physical network devices for root container.
+		nd, err := self.GetRootNetworkDevices()
+		if err != nil {
+			return spec, err
+		}
+		spec.HasNetwork = spec.HasNetwork || len(nd) != 0
+
+		// Get memory and swap limits of the running machine
+		memLimit, err := machine.GetMachineMemoryCapacity()
+		if err != nil {
+			glog.Warningf("failed to obtain memory limit for machine container")
+			spec.HasMemory = false
+		} else {
+			spec.Memory.Limit = uint64(memLimit)
+			// Spec is marked to have memory only if the memory limit is set
+			spec.HasMemory = true
+		}
+
+		swapLimit, err := machine.GetMachineSwapCapacity()
+		if err != nil {
+			glog.Warningf("failed to obtain swap limit for machine container")
+		} else {
+			spec.Memory.SwapLimit = uint64(swapLimit)
+		}
+	}
+
+	return spec, nil
 }
 
 func (self *rawContainerHandler) getFsStats(stats *info.ContainerStats) error {
@@ -259,28 +269,7 @@ func (self *rawContainerHandler) GetContainerLabels() map[string]string {
 }
 
 func (self *rawContainerHandler) ListContainers(listType container.ListType) ([]info.ContainerReference, error) {
-	containers := make(map[string]struct{})
-	for _, cgroupPath := range self.cgroupPaths {
-		err := common.ListDirectories(cgroupPath, self.name, listType == container.ListRecursive, containers)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	// Make into container references.
-	ret := make([]info.ContainerReference, 0, len(containers))
-	for cont := range containers {
-		ret = append(ret, info.ContainerReference{
-			Name: cont,
-		})
-	}
-
-	return ret, nil
-}
-
-func (self *rawContainerHandler) ListThreads(listType container.ListType) ([]int, error) {
-	// TODO(vmarmol): Implement
-	return nil, nil
+	return common.ListContainers(self.name, self.cgroupPaths, listType)
 }
 
 func (self *rawContainerHandler) ListProcesses(listType container.ListType) ([]int, error) {
@@ -433,11 +422,5 @@ func (self *rawContainerHandler) StopWatchingSubcontainers() error {
 }
 
 func (self *rawContainerHandler) Exists() bool {
-	// If any cgroup exists, the container is still alive.
-	for _, cgroupPath := range self.cgroupPaths {
-		if utils.FileExists(cgroupPath) {
-			return true
-		}
-	}
-	return false
+	return common.CgroupExists(self.cgroupPaths)
 }